feat: Limit user access to set up features
Belongs to #50.
User access has been limited to muscle groups and exercises.
A user can only view Muscle Groups and Exercises created by themselves
or the admin user.
A user can only edit and delete Muscle Groups and Exercises created by
themselves.

+++ setup/views.py
@@ -4,7 +4,6 @@ This module specifies the views used in the setup app.
 from django.shortcuts import render, redirect, get_object_or_404
 from django.views import generic
 from django.contrib import messages
-from django.contrib.auth.models import User
 from django.contrib.auth.decorators import login_required
 from .models import Exercise, MuscleGroup
 from .forms import ExerciseForm, MuscleGroupForm
@@ -12,11 +11,17 @@ from .forms import ExerciseForm, MuscleGroupForm
 
 class ExerciseList(generic.ListView):
     """
-    Renders a list of all Exercises
+    Renders a list of the users (and admins) all Exercises
     """
     model = Exercise
     template_name = 'setup/exercises.html'
     paginate_by = 50
+
+    def get_queryset(self):
+        queryset = super().get_queryset()
+        filter = [self.request.user.username, "itrainadmin"]
+        queryset = queryset.filter(created_by__username__in=filter)
+        return queryset
 
 
 @login_required
@@ -24,12 +29,11 @@ def create_exercise(request):
     """
     For creating a new exercise
     """
-    user = get_object_or_404(User, username=request.user)
+    user = request.user
     if request.method == 'POST':
-        exercise_form = ExerciseForm(request.POST)
+        exercise_form = ExerciseForm(request.POST, user=user)
         if exercise_form.is_valid():
             exercise_form.instance.created_by = user
-            exercise_form.save()
             exercise = exercise_form.save()
             messages.add_message(
                 request,
@@ -37,7 +41,7 @@ def create_exercise(request):
                 f'{exercise.name} was successfully created'
             )
             return redirect('exercises')
-    exercise_form = ExerciseForm()
+    exercise_form = ExerciseForm(user=user)
     context = {
         'form': exercise_form
     }
@@ -50,8 +54,17 @@ def edit_exercise(request, exercise_id):
     For editing a specific exercise picked from the exercise list
     """
     exercise = get_object_or_404(Exercise, id=exercise_id)
+    user = request.user
+    if user.username != exercise.created_by.username:
+        messages.add_message(
+            request,
+            messages.ERROR,
+            f"You don't have permissions to edit {exercise.name}"
+        )
+        return redirect('exercises')
     if request.method == 'POST':
-        exercise_form = ExerciseForm(request.POST, instance=exercise)
+        exercise_form = ExerciseForm(request.POST, instance=exercise,
+                                     user=user)
         if exercise_form.is_valid():
             exercise = exercise_form.save()
             messages.add_message(
@@ -60,7 +73,7 @@ def edit_exercise(request, exercise_id):
                 f'{exercise.name} was successfully saved'
             )
             return redirect('exercises')
-    exercise_form = ExerciseForm(instance=exercise)
+    exercise_form = ExerciseForm(instance=exercise, user=user)
     context = {
         'form': exercise_form,
         'exercise': exercise,
@@ -80,11 +93,17 @@ def delete_exercise(request, exercise_id):
 
 class MuscleGroupList(generic.ListView):
     """
-    Renders a list of all muscle groups
+    Renders a list of the users (and admins) all muscle groups
     """
     model = MuscleGroup
     template_name = 'setup/muscle_groups.html'
     paginate_by = 50
+
+    def get_queryset(self):
+        queryset = super().get_queryset()
+        filter = [self.request.user.username, "itrainadmin"]
+        queryset = queryset.filter(created_by__username__in=filter)
+        return queryset
 
 
 @login_required
@@ -92,9 +111,9 @@ def create_muscle_group(request):
     """
     For creating a new muscle group
     """
-    user = get_object_or_404(User, username=request.user)
+    user = request.user
     if request.method == 'POST':
-        muscle_group_form = MuscleGroupForm(request.POST)
+        muscle_group_form = MuscleGroupForm(request.POST, user=user)
         if muscle_group_form.is_valid():
             muscle_group_form.instance.created_by = user
             muscle_group = muscle_group_form.save()
@@ -104,7 +123,7 @@ def create_muscle_group(request):
                 f'{muscle_group.name} was successfully created'
             )
             return redirect('muscle_groups')
-    muscle_group_form = MuscleGroupForm()
+    muscle_group_form = MuscleGroupForm(user=user)
     context = {
         'form': muscle_group_form
     }
@@ -117,6 +136,14 @@ def edit_muscle_group(request, muscle_group_id):
     For editing a specific muscle group picked from the muscle group list
     """
     muscle_group = get_object_or_404(MuscleGroup, id=muscle_group_id)
+    user = request.user
+    if user.username != muscle_group.created_by.username:
+        messages.add_message(
+            request,
+            messages.ERROR,
+            f"You don't have permissions to edit {muscle_group.name}"
+        )
+        return redirect('muscle_groups')
     if request.method == 'POST':
         muscle_group_form = MuscleGroupForm(request.POST,
                                             instance=muscle_group)
@@ -142,5 +169,13 @@ def delete_muscle_group(request, muscle_group_id):
     For deleting a specific muscle group picked from the muscle group list
     """
     muscle_group = get_object_or_404(MuscleGroup, id=muscle_group_id)
+    user = request.user
+    if user.username != muscle_group.created_by.username:
+        messages.add_message(
+            request,
+            messages.ERROR,
+            f"You don't have permissions to delete {muscle_group.name}"
+        )
+        return redirect('muscle_groups')
     muscle_group.delete()
     return redirect('muscle_groups')
